Add all authorization

+++ api/Product/delete.php
@@ -1,4 +1,6 @@
 <?php
+  session_start();
+
   header('Access-Control-Allow-Origin: *');
   header('Content-Type: application/json');
   header('Access-Control-Allow-Methods: DELETE');
@@ -7,19 +9,27 @@
   include_once '../../config/Database.php';
   include_once '../../model/Product.php';
 
-  $database = new Database();
-  $db = $database->connect();
+  if (isset($_SESSION['user_id'])
+    && isset($_SESSION['user_type'])
+    && $_SESSION['user_type'] == 'admin'
+  ) {
+    $database = new Database();
+    $db = $database->connect();
 
-  $product = new Product($db);
-  $data = json_decode(file_get_contents("php://input"));
-  $product->product_id = $data->product_id;
+    $product = new Product($db);
+    $data = json_decode(file_get_contents("php://input"));
+    $product->product_id = $data->product_id;
 
-  if($product->delete()) {
-    echo json_encode(
-      array('message' => 'Product Deleted')
-    );
+    if($product->delete()) {
+      echo json_encode(
+        array('message' => 'Product Deleted')
+      );
+    } else {
+      echo json_encode(
+        array('message' => 'Product Not Deleted')
+      );
+    }
   } else {
-    echo json_encode(
-      array('message' => 'Product Not Deleted')
-    );
+    http_response_code(401);
+    echo json_encode(array('message' => 'You Are Not Authorized'));
   }
