+++ api/PurchasedBy/delete.php
@@ -1,4 +1,6 @@
 <?php
+  session_start();
+
   header('Access-Control-Allow-Origin: *');
   header('Content-Type: application/json');
   header('Access-Control-Allow-Methods: DELETE');
@@ -7,21 +9,29 @@
   include_once '../../config/Database.php';
   include_once '../../model/PurchasedBy.php';
 
-  $database = new Database();
-  $db = $database->connect();
+  if (isset($_SESSION['user_id'])
+    && isset($_SESSION['user_type'])
+    && $_SESSION['user_type'] == 'admin'
+  ) {
+    $database = new Database();
+    $db = $database->connect();
 
-  $pb = new PurchasedBy($db);
-  $data = json_decode(file_get_contents("php://input"));
+    $pb = new PurchasedBy($db);
+    $data = json_decode(file_get_contents("php://input"));
 
-  $pb->product_id = $data->product_id;
-  $pb->user_id = $data->user_id;
+    $pb->product_id = $data->product_id;
+    $pb->user_id = $data->user_id;
 
-  if($pb->delete()) {
-    echo json_encode(
-      array('message' => 'Purchased By Deleted')
-    );
+    if($pb->delete()) {
+      echo json_encode(
+        array('message' => 'Purchased By Deleted')
+      );
+    } else {
+      echo json_encode(
+        array('message' => 'Purchased By Not Deleted')
+      );
+    }
   } else {
-    echo json_encode(
-      array('message' => 'Purchased By Not Deleted')
-    );
+    http_response_code(401);
+    echo json_encode(array('message' => 'You Are Not Authorized'));
   }
