+++ api/Receipt/delete.php
@@ -1,4 +1,6 @@
 <?php
+  session_start();
+
   header('Access-Control-Allow-Origin: *');
   header('Content-Type: application/json');
   header('Access-Control-Allow-Methods: DELETE');
@@ -7,20 +9,28 @@
   include_once '../../config/Database.php';
   include_once '../../model/Receipt.php';
 
-  $database = new Database();
-  $db = $database->connect();
-	$receipt = new Receipt($db);
+  if (isset($_SESSION['user_id'])
+    && isset($_SESSION['user_type'])
+    && $_SESSION['user_type'] == 'admin'
+  ) {
+    $database = new Database();
+    $db = $database->connect();
+    $receipt = new Receipt($db);
 
-  $data = json_decode(file_get_contents("php://input"));
+    $data = json_decode(file_get_contents("php://input"));
 
-  $receipt->number = $data->number;
+    $receipt->number = $data->number;
 
-  if ($receipt->delete()) {
-    echo json_encode(
-      array('message' => 'Receipt Deleted')
-    );
+    if ($receipt->delete()) {
+      echo json_encode(
+        array('message' => 'Receipt Deleted')
+      );
+    } else {
+      echo json_encode(
+        array('message' => 'Receipt Not Deleted')
+      );
+    }
   } else {
-    echo json_encode(
-      array('message' => 'Receipt Not Deleted')
-    );
+    http_response_code(401);
+    echo json_encode(array('message' => 'You Are Not Authorized'));
   }
